feat(clans): refactor public join functionality into clanController and remove deprecated controller
